chore: Update writeCID method to accept Union[str, Path] for filePath
This commit modifies the `writeCID` method in the `Web3` class to accept a `Union[str, Path]` type for the `filePath` parameter. This allows the method to handle both string paths and `Path` objects. The code checks the type of `filePath` and converts it to a `Path` object if necessary before writing the CID contents to the specified file path. This update improves the flexibility and usability of the `writeCID` method.